ppc64: Fix broken SYS V shared memory support
The initial support of the SYS V shared memory on ppc64 is broken. The call
to shmat done in the restore blob has no chance to work correctly.

This patch fixes the sys_shmat call.

+++ arch/ppc64/include/asm/restorer.h
@@ -131,6 +131,6 @@ static inline int sigreturn_prep_fpu_frame(struct rt_sigframe *sigframe,
 /*
  * Defined in arch/ppc64/syscall-common-ppc64.S
  */
-int sys_shmat(int shmid, const void *shmaddr, int shmflg);
+unsigned long sys_shmat(int shmid, const void *shmaddr, int shmflg);
 
 #endif /*__CR_ASM_RESTORER_H__*/
